backend/requirements.txt
langchain>=0.3.25
langchain-core>=0.3.25
langchain-experimental>=0.3.4
langchain-mcp-adapters>=0.1.1
langchain-ollama>=0.3.2
langchain-openai>=0.3.16
langgraph>=0.4.8
langgraph-checkpoint-mongodb>=0.1.4
langgraph-checkpoint-redis>=0.0.7

openai>=1.78.0
pydantic>=2.11.4

requests>=2.32.0
python-dotenv>=1.0.0
mcp>=1.9.1

redis>=5.0.0
pymongo>=4.0.0
motor>=3.7.0

click>=8.2.0
typing-extensions>=4.0.0 

pyautogui
psutil